Refactoring + fix test
