+++ resources/logback.xml
@@ -15,7 +15,7 @@
 
 <!-- <logger name="main.java" level="DEBUG"/> -->
 
-<root level="WARN">
+<root level="INFO">
 	<appender-ref ref="FILE" />
 	<appender-ref ref="STDOUT" />
 </root>
