+++ src/main/java/distributed/results/WebsiteResult.java
@@ -24,11 +24,46 @@ import com.j256.ormlite.table.*;
 	@DatabaseField
 	private final float duration;
 	
-	@ForeignCollectionField(eager = false)
+	@ForeignCollectionField(eager = true)
     private ForeignCollection<StateResult> stateResults;
 	
 	public WebsiteResult() {
 		jsonResults = null;
 		duration = 0;
 	}
+
+	@Override
+	public boolean equals(Object obj) {
+		if (this == obj)
+			return true;
+		if (obj == null)
+			return false;
+		if (getClass() != obj.getClass())
+			return false;
+		WebsiteResult other = (WebsiteResult) obj;
+		if (Float.floatToIntBits(duration) != Float
+				.floatToIntBits(other.duration))
+			return false;
+		if (id != other.id)
+			return false;
+		if (jsonResults == null) {
+			if (other.jsonResults != null)
+				return false;
+		} else if (!jsonResults.equals(other.jsonResults))
+			return false;
+		return true;
+	}
+
+	@Override
+	public int hashCode() {
+		final int prime = 31;
+		int result = 1;
+		result = prime * result + Float.floatToIntBits(duration);
+		result = prime * result + id;
+		result = prime * result
+				+ ((jsonResults == null) ? 0 : jsonResults.hashCode());
+		return result;
+	}
+	
+	
 }
